Move column choice types to separate module
