+++ src/features/views/components/ViewColumnDialog/choices/index.tsx
@@ -1,50 +1,9 @@
-import { IntlShape } from 'react-intl';
-
-import {
-  PendingZetkinViewColumn,
-  SelectedViewColumn,
-  ZetkinViewColumn,
-} from '../../types';
+import { CHOICES, ColumnChoice, ColumnChoiceWithKey } from './types';
 
 import * as fields from './fields';
 import * as misc from './misc';
 import * as query from './query';
 import * as surveys from './surveys';
-
-export enum CHOICES {
-  DELEGATE = 'delegate',
-  FULL_NAME = 'fullName',
-  FIRST_NAME = 'firstName',
-  LAST_NAME = 'lastName',
-  EMAIL = 'email',
-  PHONE = 'phone',
-  FOLLOW_UP = 'followUp',
-  PERSON_FIELDS = 'personFields',
-  SURVEY_SUBMIT_DATE = 'surveySubmitDate',
-  TAG = 'tag',
-  TOGGLE = 'toggle',
-  ASSIGNEE = 'localPerson',
-  NOTES = 'localText',
-  CUSTOM_QUERY = 'localQuery',
-  SURVEY_RESPONSE = 'surveyResponse',
-  SURVEY_RESPONSES = 'surveyResponses',
-}
-
-export type ColumnChoice = {
-  alreadyInView?: (columns: ZetkinViewColumn[]) => boolean;
-  color: string;
-  defaultColumns?: (
-    intl: IntlShape,
-    columns: ZetkinViewColumn[]
-  ) => PendingZetkinViewColumn[];
-  renderCardVisual: (color: string) => JSX.Element;
-  renderConfigForm?: (props: {
-    existingColumns: ZetkinViewColumn[];
-    onOutputConfigured: (columns: SelectedViewColumn[]) => void;
-  }) => JSX.Element;
-};
-
-export type ColumnChoiceWithKey = ColumnChoice & { key: CHOICES };
 
 const choices: Record<CHOICES, ColumnChoice> = {
   [CHOICES.FULL_NAME]: fields.fullName,
@@ -66,3 +25,5 @@ const choices: Record<CHOICES, ColumnChoice> = {
 };
 
 export default choices;
+
+export type { CHOICES, ColumnChoice, ColumnChoiceWithKey };
